+++ BaseballLineupIOS/View/ViewController/OrderViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import GoogleMobileAds
 
-class OrderViewController: UIViewController {
+class OrderViewController: ADBaseViewController {
     
     var viewModel: OrderViewModel?
     var parentViewModel: CustomTabBarViewModel?
@@ -82,29 +82,8 @@ class OrderViewController: UIViewController {
         bannerAD.rootViewController = self
     }
     
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        loadBannerAd()
-    }
-    
-    override func viewWillTransition(to size: CGSize,
-                                     with coordinator: UIViewControllerTransitionCoordinator) {
-        super.viewWillTransition(to:size, with:coordinator)
-        coordinator.animate(alongsideTransition: { _ in
-            self.loadBannerAd()
-        })
-    }
-    
-    func loadBannerAd() {
-        let frame = { () -> CGRect in
-            if #available(iOS 11.0, *) {
-                return view.frame.inset(by: view.safeAreaInsets)
-            } else {
-                return view.frame
-            }
-        }()
-        let viewWidth = frame.size.width
-        bannerAD.adSize = GADCurrentOrientationAnchoredAdaptiveBannerAdSizeWithWidth(viewWidth)
+    override func loadBannerAd() {
+        bannerAD.adSize = GADCurrentOrientationAnchoredAdaptiveBannerAdSizeWithWidth(getViewWidth())
         bannerAD.load(GADRequest())
     }
