+++ BaseballLineupIOS/View/ViewController/SubMemberViewController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 import GoogleMobileAds
-class SubMemberViewController: UIViewController {
+class SubMemberViewController: ADBaseViewController {
     
     var viewModel: SubMemberViewModel?
     var parentViewModel: CustomTabBarViewModel?
@@ -96,31 +96,13 @@ class SubMemberViewController: UIViewController {
         nameTF.delegate = self
         viewModel?.delegate = self
         setDefaultUIState()
+        
+        bannerAD.adUnitID = Constants.BANNER_ID
+        bannerAD.rootViewController = self
     }
     
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        loadBannerAd()
-    }
-    
-    override func viewWillTransition(to size: CGSize,
-                                     with coordinator: UIViewControllerTransitionCoordinator) {
-        super.viewWillTransition(to:size, with:coordinator)
-        coordinator.animate(alongsideTransition: { _ in
-            self.loadBannerAd()
-        })
-    }
-    
-    func loadBannerAd() {
-        let frame = { () -> CGRect in
-            if #available(iOS 11.0, *) {
-                return view.frame.inset(by: view.safeAreaInsets)
-            } else {
-                return view.frame
-            }
-        }()
-        let viewWidth = frame.size.width
-        bannerAD.adSize = GADCurrentOrientationAnchoredAdaptiveBannerAdSizeWithWidth(viewWidth)
+    override func loadBannerAd() {
+        bannerAD.adSize = GADCurrentOrientationAnchoredAdaptiveBannerAdSizeWithWidth(getViewWidth())
         bannerAD.load(GADRequest())
     }
